Added client and manager pages gui

+++ lab_07/src/business_logic/models/ModelClient.h
@@ -5,7 +5,7 @@
 #include <iostream>
 #include <string>
 
-typedef int passport_t;
+typedef long int passport_t;
 
 class Client {
 private:
